v3.0.177

+++ src/Listener/MainWorkerStartListener.php
@@ -15,11 +15,14 @@ use GuzzleHttp\Client;
 use GuzzleHttp\Exception\GuzzleException;
 use Hyperf\Amqp\Annotation\Consumer;
 use Hyperf\Amqp\Annotation\Producer;
+use Hyperf\Contract\ApplicationInterface;
 use Hyperf\Di\Annotation\AnnotationCollector;
 use Hyperf\Event\Annotation\Listener;
 use Hyperf\Event\Contract\ListenerInterface;
 use Hyperf\Framework\Event\MainWorkerStart;
 use Swoole\Process;
+use Symfony\Component\Console\Input\ArrayInput;
+use Symfony\Component\Console\Output\NullOutput;
 
 /**
  * Hyperf worker 启动后执行.
@@ -36,27 +39,41 @@ class MainWorkerStartListener implements ListenerInterface
 
     public function process(object $event): void
     {
+        // 生产环境，执行下 preStart，初始下sql语句
+        if (env('APP_ENV') == 'production') {
+            $input = new ArrayInput(['command' => 'preStart']);
+            $output = new NullOutput();
+            $application = container()->get(ApplicationInterface::class);
+            $application->setAutoExit(false);
+            $exitCode = $application->run($input, $output);
+            stdLog()->info('preStart result：', [$exitCode]);
+        }
+
         // 检测mq的queue、exchange是否以当前服务名开始，避免复制其他代码导致queue相同，引发问题
         if (env('AMQP_USER') && env('AMQP_PASSWORD') && env('APP_NAME')) {
             $consumerExchangeArr = [];
             // Consumer的queue必须以当前服务名开始
             $class = AnnotationCollector::getClassesByAnnotation(Consumer::class);
-            foreach ($class as $item) {
-                if (! empty($item->queue) && stripos($item->queue, env('APP_NAME')) !== 0) {
-                    stdLog()->error('发现mq的queue不符合规则，必须以服务名（' . env('APP_NAME') . '）开始：' . $item->queue);
-                    Process::kill((int) file_get_contents(\Hyperf\Config\config('server.settings.pid_file')));
-                    break;
+            if (! empty($class)) {
+                foreach ($class as $item) {
+                    if (! empty($item->queue) && stripos($item->queue, env('APP_NAME')) !== 0) {
+                        stdLog()->error('发现mq消费者的queue不符合规则，必须以服务名（' . env('APP_NAME') . '）开始：' . $item->queue);
+                        Process::kill((int) file_get_contents(\Hyperf\Config\config('server.settings.pid_file')));
+                        break;
+                    }
+                    $consumerExchangeArr[] = $item->exchange;
                 }
-                $consumerExchangeArr[] = $item->exchange;
             }
 
             // Producer的exchange必须要以本服务名开始，特别是当本服务的Consumer存在的时候，避免命令为其他服务。
             $class = AnnotationCollector::getClassesByAnnotation(Producer::class);
-            foreach ($class as $item) {
-                if (! empty($item->exchange) && stripos($item->exchange, env('APP_NAME')) !== 0 && in_array($item->exchange, $consumerExchangeArr)) {
-                    stdLog()->error('发现mq的exchange不符合规则，必须以服务名（' . env('APP_NAME') . '）开始：' . $item->exchange);
-                    Process::kill((int) file_get_contents(\Hyperf\Config\config('server.settings.pid_file')));
-                    break;
+            if (! empty($class)) {
+                foreach ($class as $item) {
+                    if (! empty($item->exchange) && stripos($item->exchange, env('APP_NAME')) !== 0 && in_array($item->exchange, $consumerExchangeArr)) {
+                        stdLog()->error('发现mq投递者的exchange不符合规则，必须以服务名（' . env('APP_NAME') . '）开始：' . $item->exchange);
+                        Process::kill((int) file_get_contents(\Hyperf\Config\config('server.settings.pid_file')));
+                        break;
+                    }
                 }
             }
         }
